Re write chase-lev
